requirements.txt
Flask==2.3.2
Flask-Bcrypt==1.0.1
Flask-JWT-Extended==4.5.2
Flask-SQLAlchemy==3.0.5
openai==0.27.8
pdfplumber==0.10.2
psycopg2-binary==2.9.7
python-dotenv==1.0.0
regex==2023.8.8
